Removed redundant antialiased waveform types and moved antialiasing into audio oscillator.
