+++ src/audio/signal/waveform/Pulse.php
@@ -24,28 +24,106 @@ use ABadCafe\PDE\Util;
 use function \ceil;
 
 /**
- * Pulse
+ * AliasedPulse
  *
- * PWM implementation of IWaveform
+ * PWM implementation of IWaveform. Does not attempt to peform anti aliasing on the output. Faster for low frequencies
+ * and interestingly noisy for higher frequencies.
  *
  * @see https://github.com/0xABADCAFE/random-proto-synth
  */
-class Pulse extends AliasedPulse {
+class Pulse implements IHardTransient {
 
-    use TAntialiased;
+    const
+        /**
+         * Waveform period (interval after which it repeats).
+         */
+        PERIOD    = 1.0,
+        MIN_WIDTH = 1.0/16.0,
+        MAX_WIDTH = 15.0/16.0,
+        DEF_WIDTH = 0.25
+    ;
+
+    protected float $fPulseWidth = 0.25;
+
+    /**
+     * Optional modulator for pulsewidth.
+     */
+    protected ?Signal\IStream $oWidthModulator = null;
+
+    /**
+     * Constructor.
+     *
+     * @param float               $fPulseWidth - how far into the duty cycle the switch happens.
+     * @param Signal\IStream|null $oModulator  - optional modulator for pulse width
+     */
+    public function __construct(float $fPulseWidth = self::DEF_WIDTH, ?Signal\IStream $oModulator = null) {
+        $this->setPulsewidth($fPulseWidth);
+        $this->setPulsewidthModulator($oModulator);
+    }
+
+    public function __clone(): void {
+        if ($this->oWidthModulator) {
+            if ($this->oWidthModulator instanceof Util\ISometimesShareable) {
+                $this->oWidthModulator = $this->oWidthModulator->share(); // @phpstan-ignore-line
+            } else {
+                $this->oWidthModulator = clone $this->oWidthModulator;
+            }
+        }
+    }
+
+    /**
+     * @inheritDoc
+     *
+     * If there is no modulator attached, then the instance can be shared idempotently.
+     */
+    public function share(): self {
+        return $this->oWidthModulator ? clone $this : $this;
+    }
+
+    /**
+     * Set the pulse width. Clamps to MIN_WIDTH/MAX_WIDTH
+     *
+     * @param  float $fPulseWidth
+     * @return self
+     */
+    public function setPulsewidth(float $fPulseWidth): self {
+        if ($fPulseWidth > self::MAX_WIDTH) {
+            $fPulseWidth = self::MAX_WIDTH;
+        }
+        if ($fPulseWidth < self::MIN_WIDTH) {
+            $fPulseWidth = self::MIN_WIDTH;
+        }
+        $this->fPulseWidth = $fPulseWidth;
+        return $this;
+    }
+
+    /**
+     * Optional pulse width modulator to use. Note that this creates free-running clone of the input.
+     *
+     * @param  Signal\IStream|null $oModulator
+     * @return self
+     */
+    public function setPulsewidthModulator(?Signal\IStream $oModulator): self {
+        if ($oModulator) {
+            $this->oWidthModulator = clone $oModulator;
+        } else {
+            $this->oWidthModulator = null;
+        }
+        return $this;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getPeriod(): float {
+        return self::PERIOD;
+    }
 
     /**
      * @inheritDoc
      */
     public function map(Signal\Packet $oInput): Signal\Packet {
         $oOutput = clone $oInput;
-
-        // Avoid sharp transitions at the edges with a simple hamming filter.
-        $fPrev1  = $this->fPrev1;
-        $fPrev2  = $this->fPrev2;
-        $fPrev3  = $this->fPrev3;
-        $fPrev4  = $this->fPrev4;
-
         if ($this->oWidthModulator) {
             $oWidth = $this->oWidthModulator
                 ->emit()
@@ -53,38 +131,21 @@ class Pulse extends AliasedPulse {
                 ->biasBy(0.5);
             foreach ($oInput as $i => $fTime) {
                 /** @var float $fTime */
-                $fSample = ((ceil($fTime) - $fTime) > $oWidth[$i]) ? 1.0 : -1.0;
-                $oOutput[$i] = 0.1 * (
-                    $fSample + $fPrev4 +
-                    2.0 * ($fPrev1 + $fPrev3)
-                    + 4.0 * $fPrev2
-                );
-                $fPrev4 = $fPrev3;
-                $fPrev3 = $fPrev2;
-                $fPrev2 = $fPrev1;
-                $fPrev1 = $fSample;
+                $oOutput[$i] = ((ceil($fTime) - $fTime) > $oWidth[$i]) ? 1.0 : -1.0;
             }
         } else {
             foreach ($oInput as $i => $fTime) {
                 /** @var float $fTime */
-                $fSample = ((ceil($fTime) - $fTime) > $this->fPulseWidth) ? 1.0 : -1.0;
-                $oOutput[$i] = 0.1 * (
-                    $fSample + $fPrev4 +
-                    2.0 * ($fPrev1 + $fPrev3)
-                    + 4.0 * $fPrev2
-                );
-                $fPrev4 = $fPrev3;
-                $fPrev3 = $fPrev2;
-                $fPrev2 = $fPrev1;
-                $fPrev1 = $fSample;
+                $oOutput[$i] = ((ceil($fTime) - $fTime) > $this->fPulseWidth) ? 1.0 : -1.0;
             }
         }
-
-        $this->fPrev1 = $fPrev1;
-        $this->fPrev2 = $fPrev2;
-        $this->fPrev3 = $fPrev3;
-        $this->fPrev4 = $fPrev4;
-
         return $oOutput;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function value(float $fInput): float {
+        return ((ceil($fTime) - $fTime) > $this->fPulseWidth) ? 1.0 : -1.0;
     }
 }
