+++ src/audio/signal/waveform/Square.php
@@ -24,43 +24,41 @@ use ABadCafe\PDE\Util;
 use function \floor;
 
 /**
- * Square
+ * AliasedSquare
  *
- * Square implementation of IWaveform
+ * Square implementation of IWaveform.
  *
  * @see https://github.com/0xABADCAFE/random-proto-synth
  */
-class Square extends AliasedSquare {
+class Square implements IHardTransient {
 
-    use TAntialiased;
+    use Util\TAlwaysShareable;
+
+    /**
+     * Waveform period (interval after which it repeats).
+     */
+    const PERIOD = 2.0;
+
+    /**
+     * @inheritDoc
+     */
+    public function getPeriod(): float {
+        return self::PERIOD;
+    }
 
     /**
      * @inheritDoc
      */
     public function map(Signal\Packet $oInput): Signal\Packet {
         $oOutput = clone $oInput;
-
-        // Avoid sharp transitions at the edges with a simple hamming filter.
-        $fPrev1  = $this->fPrev1;
-        $fPrev2  = $this->fPrev2;
-        $fPrev3  = $this->fPrev3;
-        $fPrev4  = $this->fPrev4;
         foreach ($oInput as $i => $fTime) {
-            $fSample = (int)floor($fTime) & 1 ? -1.0 : 1.0; // @phpstan-ignore-line - false positive
-            $oOutput[$i] = 0.1 * (
-                $fSample + $fPrev4 +
-                2.0 * ($fPrev1 + $fPrev3)
-                + 4.0 * $fPrev2
-            );
-            $fPrev4 = $fPrev3;
-            $fPrev3 = $fPrev2;
-            $fPrev2 = $fPrev1;
-            $fPrev1 = $fSample;
+            /** @var float $fTime */
+            $oOutput[$i] = (int)floor($fTime) & 1 ? -1.0 : 1.0;
         }
-        $this->fPrev1 = $fPrev1;
-        $this->fPrev2 = $fPrev2;
-        $this->fPrev3 = $fPrev3;
-        $this->fPrev4 = $fPrev4;
         return $oOutput;
+    }
+
+    public function value(float $fTime): float {
+        return (int)floor($fTime) & 1 ? -1.0 : 1.0;
     }
 }
