--- src/audio/signal/waveform/TAntialiased.php
@@ -1,37 +1,0 @@
-<?php
-/**
- *                   ______                            __
- *           __     /\\\\\\\\_                        /\\\
- *          /\\\  /\\\//////\\\_                      \/\\\
- *        /\\\//  \///     \//\\\    ________       ___\/\\\         _______
- *      /\\\//               /\\\   /\\\\\\\\\_    /\\\\\\\\\       /\\\\\\\\_
- *    /\\\//_              /\\\\/   /\\\/////\\\   /\\\////\\\     /\\\/////\\\
- *    \////\\\ __          /\\\/    \/\\\   \/\\\  \/\\\  \/\\\    /\\\\\\\\\\\
- *        \////\\\ __      \///_     \/\\\___\/\\\  \/\\\__\/\\\   \//\\\//////_
- *            \////\\\       /\\\     \/\\\\\\\\\\   \//\\\\\\\\\    \//\\\\\\\\\
- *                \///       \///      \/\\\//////     \/////////      \/////////
- *                                      \/\\\
- *                                       \///
- *
- *                         /P(?:ointless|ortable|HP) Demo Engine/
- */
-
-declare(strict_types=1);
-
-namespace ABadCafe\PDE\Audio\Signal\Waveform;
-use ABadCafe\PDE\Util;
-
-/**
- * Antialiased using a filter window
- */
-trait TAntialiased {
-
-    use Util\TNeverShareable;
-
-    private float
-        $fPrev1 = 0.0,
-        $fPrev2 = 0.0,
-        $fPrev3 = 0.0,
-        $fPrev4 = 0.0
-    ;
-}
