+++ src/audio/signal/waveform/sine/SineXForm.php
@@ -96,6 +96,12 @@ abstract class SineXForm implements Signal\IWaveform {
     }
 
     public function value(float $fInput): float {
-        return sin($fInput);
+        $aTransform = static::TRANSFORM;
+        $fPhase     = $fInput;
+        $iQuadrant  = ((int)$fPhase) & 3;
+        $fPhase     += $aTransform[$iQuadrant][0];
+        $fPhase     *= self::ADJUST;
+        $fSin       = sin($fPhase);
+        return ($fSin * $aTransform[$iQuadrant][2]) + $aTransform[$iQuadrant][1];
     }
 }
