+++ src/classmap.php
@@ -77,16 +77,13 @@ const CLASS_MAP = [
   'ABadCafe\\PDE\\Audio\\Signal\\Oscillator\\LFO' => '/audio/signal/oscillator/LFO.php',
   'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\Triangle' => '/audio/signal/waveform/Triangle.php',
   'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\Square' => '/audio/signal/waveform/Square.php',
-  'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\AliasedSquare' => '/audio/signal/waveform/AliasedSquare.php',
-  'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\AliasedPulse' => '/audio/signal/waveform/AliasedPulse.php',
   'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\TriangleHalfRect' => '/audio/signal/waveform/TriangleHalfRect.php',
   'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\WhiteNoise' => '/audio/signal/waveform/WhiteNoise.php',
-  'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\TAntialiased' => '/audio/signal/waveform/TAntialiased.php',
   'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\Pulse' => '/audio/signal/waveform/Pulse.php',
-  'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\AliasedSaw' => '/audio/signal/waveform/AliasedSaw.php',
   'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\Factory' => '/audio/signal/waveform/Factory.php',
   'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\Flyweight' => '/audio/signal/waveform/Flyweight.php',
   'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\Saw' => '/audio/signal/waveform/Saw.php',
+  'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\IHardTransient' => '/audio/signal/waveform/IHardTransient.php',
   'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\SineHalfRect' => '/audio/signal/waveform/sine/SineHalfRect.php',
   'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\Sine' => '/audio/signal/waveform/sine/Sine.php',
   'ABadCafe\\PDE\\Audio\\Signal\\Waveform\\SineCut' => '/audio/signal/waveform/sine/SineCut.php',
