+++ src/tests/test_oscillator_feedback.php
@@ -6,7 +6,12 @@ namespace ABadCafe\PDE;
 
 require_once '../PDE.php';
 
-const F_INDEX = 0.75;
+const F_INDEX = 1.0;
+
+$oModulator = new Audio\Signal\Oscillator\Sound(
+    Audio\Signal\Waveform\Flyweight::get()->getWaveform(Audio\Signal\IWaveform::TRIANGLE),
+    220.0
+);
 
 $oOscillator = new Audio\Signal\Oscillator\Sound(
     Audio\Signal\Waveform\Flyweight::get()->getWaveform(Audio\Signal\IWaveform::SINE),
@@ -16,11 +21,15 @@ $oOscillator = new Audio\Signal\Oscillator\Sound(
 $oEnvelope = new Audio\Signal\Envelope\Shape(
     0.0,
     [
-        [1.0, 1.0]
+        [1.0, 1.0],
+        [1.0, 0.25],
+        [0.0, 2]
     ]
 );
 
 $oOscillator
+    //->setPhaseModulator($oModulator)
+    //->setPhaseModulationIndex(0.5)
     ->setLevelEnvelope($oEnvelope)
     ->setPhaseFeedbackIndex(F_INDEX);
